+++ src/components/ChannelList.js
@@ -119,6 +119,21 @@ const ChannelList = withChatContext(
        * If true, channels won't be dynamically sorted by most recent message.
        */
       lockChannelOrder: PropTypes.bool,
+      /**
+       * Besides existing (default) UX behaviour of underlying flatlist of ChannelList component, if you want
+       * to attach some additional props to un derlying flatlist, you can add it to following prop.
+       *
+       * You can find list of all the available FlatList props here - https://facebook.github.io/react-native/docs/flatlist#props
+       *
+       * e.g.
+       * ```
+       * <ChannelList
+       *  filters={filters}
+       *  sort={sort}
+       *  additionalFlatListProps={{ bounces: true }} />
+       * ```
+       */
+      additionalFlatListProps: PropTypes.object,
     };
 
     static defaultProps = {
@@ -133,6 +148,7 @@ const ChannelList = withChatContext(
       // https://github.com/facebook/react-native/blob/a7a7970e543959e9db5281914d5f132beb01db8d/Libraries/Lists/VirtualizedList.js#L466
       loadMoreThreshold: 2,
       lockChannelOrder: false,
+      additionalFlatListProps: {},
       logger: () => {},
     };