+++ src/components/ChannelListMessenger.js
@@ -51,6 +51,20 @@ const ChannelListMessenger = withChatContext(
       error: PropTypes.oneOfType([PropTypes.bool, PropTypes.object]),
       /** If channels are being queries. LoadingIndicator will be displayed if true */
       loadingChannels: PropTypes.bool,
+      /**
+       * Besides existing (default) UX behaviour of underlying flatlist of ChannelListMessenger component, if you want
+       * to attach some additional props to un derlying flatlist, you can add it to following prop.
+       *
+       * You can find list of all the available FlatList props here - https://facebook.github.io/react-native/docs/flatlist#props
+       *
+       * e.g.
+       * ```
+       * <ChannelListMessenger
+       *  channels={channels}
+       *  additionalFlatListProps={{ bounces: true }} />
+       * ```
+       */
+      additionalFlatListProps: PropTypes.object,
     };
 
     static defaultProps = {
@@ -60,6 +74,7 @@ const ChannelListMessenger = withChatContext(
       EmptyStateIndicator,
       // https://github.com/facebook/react-native/blob/a7a7970e543959e9db5281914d5f132beb01db8d/Libraries/Lists/VirtualizedList.js#L466
       loadMoreThreshold: 2,
+      additionalFlatListProps: {},
     };
 
     renderLoading = () => {
@@ -92,6 +107,7 @@ const ChannelListMessenger = withChatContext(
           />
         )}
         keyExtractor={(item) => item.cid}
+        {...this.props.additionalFlatListProps}
       />
     );