+++ CHANGELOG.md
@@ -7,11 +7,13 @@
 - Changes to add compatibility for Expo 36. Adding [netinfo](https://github.com/react-native-community/react-native-netinfo) as peer dependency - https://github.com/GetStream/stream-chat-react-native/issues/97
 - Disabling the longpress on image viewer (in Attachment) - https://github.com/GetStream/stream-chat-react-native/issues/100 to avoid freezing of UI.
 - Fixing broken threads issue
+- Support for `additionalFlatListProps` prop in `MessageList` and `ChannelList` component
 
 #### stream-chat-react-native (Native package)
 
 - Disabling the longpress on image viewer (in Attachment) - https://github.com/GetStream/stream-chat-react-native/issues/100 to avoid freezing of UI.
 - Fixing broken threads issue
+- Support for `additionalFlatListProps` prop in `MessageList` and `ChannelList` component
 
 ## [0.4.0] 2019-12-16